add environment print test
this can be useful when diagnosing seemingly bizarre problems
when you're not sure if the shell / frontend is passing needed
environment variables to the final executable.

For example, perhaps a Matlab/Python script calls the gemini3d.run
front-end that invokes gemini.bin. We want to be sure the final
Fortran executable has the several environment variables it needs,
particularly for MPI etc.